+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # **Upcoming release**
 
-- Remove site-packages from packages search tree (@tkrabel)
+- #733 skip directories with perm error when building autoimport index (@MrBago)
+- #722, #723 Remove site-packages from packages search tree (@tkrabel)
 
 # Release 1.11.0