+++ CHANGELOG.md
@@ -1,9 +1,14 @@
 # **Upcoming release**
+
 - #516 Autoimport Now automatically detects project dependencies and can read TOML configuration
+
+# Release 1.12.0
+
 - #733 skip directories with perm error when building autoimport index (@MrBago)
 - #722, #723 Remove site-packages from packages search tree (@tkrabel)
 - #738 Implement os.PathLike on Resource (@lieryan)
 - #739, #736 Ensure autoimport requests uses indexes (@lieryan)
+- #734, #735 raise exception when extracting the start of a block without the end
 
 # Release 1.11.0